# commander: preflight check minor optimization