--- a/src/modules/commander/Arming/PreFlightCheck/checks/distanceSensorChecks.cpp
+++ b/src/modules/commander/Arming/PreFlightCheck/checks/distanceSensorChecks.cpp
@@ -34,9 +34,9 @@
 #include "../PreFlightCheck.hpp"
 
 #include <drivers/drv_hrt.h>
+#include <HealthFlags.h>
 #include <px4_defines.h>
 #include <systemlib/mavlink_log.h>
-#include <uORB/uORB.h>
 #include <uORB/Subscription.hpp>
 #include <uORB/topics/distance_sensor.h>
 
@@ -45,27 +45,30 @@ using namespace time_literals;
 bool PreFlightCheck::distSensCheck(orb_advert_t *mavlink_log_pub, vehicle_status_s &status, const uint8_t instance,
 				   const bool is_mandatory, bool &report_fail)
 {
-	const bool exists = (orb_exists(ORB_ID(distance_sensor), instance) == PX4_OK);
-	bool valid = false;
+	uORB::SubscriptionData<distance_sensor_s> dist_sens_sub{ORB_ID(distance_sensor), instance};
+
+	const bool exists = dist_sens_sub.advertised();
+
+	bool is_required = is_mandatory;
+	bool is_valid = false;
 
 	if (exists) {
-		uORB::SubscriptionData<distance_sensor_s> dist_sens_sub{ORB_ID(distance_sensor), instance};
-		dist_sens_sub.update();
-		const distance_sensor_s &dist_sens_data = dist_sens_sub.get();
 
-		valid = (hrt_elapsed_time(&dist_sens_data.timestamp) < 1_s);
+		is_valid = (dist_sens_sub.get().timestamp != 0) && (hrt_elapsed_time(&dist_sens_sub.get().timestamp) < 1_s);
 	}
 
-	if (report_fail && is_mandatory) {
+	if (report_fail && is_required) {
 		if (!exists) {
 			mavlink_log_critical(mavlink_log_pub, "Preflight Fail: distance sensor %u missing", instance);
 			report_fail = false;
 
-		} else if (!valid) {
+		} else if (!is_valid) {
 			mavlink_log_critical(mavlink_log_pub, "Preflight Fail: no valid data from distance sensor %u", instance);
 			report_fail = false;
 		}
 	}
 
-	return valid || !is_mandatory;
+	const bool is_sensor_ok = is_valid;
+
+	return is_sensor_ok || !is_required;
 }