I was contemplating moving `commander` to a work queue and noticed that the preflight checks are quite slow. It doesn't really matter right now because we run them periodically (~ 1 Hz) in a lower priority thread, but it might in the future.

 - limit to 4 sensors
 - when iterating through estimator_status subscriptions break once they're no longer advertised
 - only proceed with calibration checks (checking all calibration slots) if the device id is populated
 - minor changes to keep things roughly aligned between accel/baro/distance/gyro/mag 

#### master
``` Console
nsh> commander status
INFO  [commander] arming: STANDBY
INFO  [commander] navigation: MANUAL
commander: cycle: 1303 events, 627197us elapsed, 481.35us avg, min 82us max 16327us 21118.385us rms
commander: preflight check: 29 events, 361609us elapsed, 12469.28us avg, min 10686us max 16208us 13505.922us rms
```


#### PR
``` Console
nsh> commander status
INFO  [commander] arming: STANDBY
INFO  [commander] navigation: MANUAL
commander: cycle: 531 events, 160720us elapsed, 302.67us avg, min 82us max 7411us 1647.625us rms
commander: preflight check: 12 events, 47280us elapsed, 3940.00us avg, min 2225us max 4872us 2372.114us rms
```