@dagar 2022-09-15
Quite a few conflicts with the new state of things, but some minor ideas here might be reused later.